poprawione

+++ src/main/java/com/managementservice/projectmanagement/repositorie/ParticipationInTheProjectRepository.java
@@ -3,6 +3,6 @@ package com.managementservice.projectmanagement.repositorie;
 import com.managementservice.projectmanagement.entity.ParticipationInTheProject;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParticipationInTheProjectRepository extends JpaRepository<ParticipationInTheProject, Integer> {
+public interface ParticipationInTheProjectRepository extends JpaRepository<ParticipationInTheProject, Long> {
 
 }
