+++ src/main/java/com/managementservice/projectmanagement/repositorie/SprintRepository.java
@@ -3,5 +3,5 @@ package com.managementservice.projectmanagement.repositorie;
 import com.managementservice.projectmanagement.entity.Sprint;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SprintRepository extends JpaRepository<Sprint, Integer> {
+public interface SprintRepository extends JpaRepository<Sprint, Long> {
 }
